Improve Radial Chart Functionality (#245)
This PR addresses a number of bugs that have been present in the radial chart for a while. Chief among them are:

When specify a variable such as opacity that happen to all have the same value, the value default to a small value. This problem is documented in #234

Radial chart not updating when new props are passed. This issue was briefly referenced in #161 but didn't have a direct gh-issue about it, so I added a example that demonstrates how to build a hover, as well demonstrates the bug.

Untangling the mouse events in radial (#157)