Pass undefined onWheel handler to avoid registering active listener when not needed (#1114)
In the case that the XYPlot is used without an onWheel callback the code was registering a listener anyway. This would cause warnings in the chrome console about registering an active listener.

The warning will still be shown in the case that an onWheel callback is set, but this change prevents the listener from being registered when it is a no-op anyway.